+++ pulsar-broker-common/src/main/java/org/apache/pulsar/broker/ServiceConfiguration.java
@@ -38,6 +38,7 @@ import org.apache.bookkeeper.client.api.DigestType;
 import org.apache.bookkeeper.util.BookKeeperConstants;
 import org.apache.commons.lang3.StringUtils;
 import org.apache.pulsar.broker.authorization.PulsarAuthorizationProvider;
+import org.apache.pulsar.broker.configuration.MetricConfiguration;
 import org.apache.pulsar.common.configuration.Category;
 import org.apache.pulsar.common.configuration.FieldContext;
 import org.apache.pulsar.common.configuration.PulsarConfiguration;
@@ -78,7 +79,8 @@ public class ServiceConfiguration implements PulsarConfiguration {
     @Category
     private static final String CATEGORY_SCHEMA = "Schema";
     @Category
-    private static final String CATEGORY_METRICS = "Metrics";
+    public static final String CATEGORY_METRICS = "Metrics";
+    private MetricConfiguration metricConfiguration;
     @Category
     private static final String CATEGORY_REPLICATION = "Replication";
     @Category
@@ -2817,89 +2819,6 @@ public class ServiceConfiguration implements PulsarConfiguration {
     )
     private int webSocketMaxTextFrameSize = 1048576;
 
-    /**** --- Metrics. --- ****/
-    @FieldContext(
-            category = CATEGORY_METRICS,
-            doc = "Whether the '/metrics' endpoint requires authentication. Defaults to false."
-                    + "'authenticationEnabled' must also be set for this to take effect."
-    )
-    private boolean authenticateMetricsEndpoint = false;
-    @FieldContext(
-        category = CATEGORY_METRICS,
-        doc = "If true, export topic level metrics otherwise namespace level"
-    )
-    private boolean exposeTopicLevelMetricsInPrometheus = true;
-    @FieldContext(
-            category = CATEGORY_METRICS,
-            doc = "If true, export buffered metrics"
-    )
-    private boolean metricsBufferResponse = false;
-    @FieldContext(
-        category = CATEGORY_METRICS,
-        doc = "If true, export consumer level metrics otherwise namespace level"
-    )
-    private boolean exposeConsumerLevelMetricsInPrometheus = false;
-    @FieldContext(
-            category = CATEGORY_METRICS,
-            doc = "If true, export producer level metrics otherwise namespace level"
-    )
-    private boolean exposeProducerLevelMetricsInPrometheus = false;
-    @FieldContext(
-            category = CATEGORY_METRICS,
-            doc = "If true, export managed ledger metrics (aggregated by namespace)"
-    )
-    private boolean exposeManagedLedgerMetricsInPrometheus = true;
-    @FieldContext(
-            category = CATEGORY_METRICS,
-            doc = "If true, export managed cursor metrics"
-    )
-    private boolean exposeManagedCursorMetricsInPrometheus = false;
-    @FieldContext(
-            category = CATEGORY_METRICS,
-            doc = "Classname of Pluggable JVM GC metrics logger that can log GC specific metrics")
-    private String jvmGCMetricsLoggerClassName;
-
-    @FieldContext(
-        category = CATEGORY_METRICS,
-        doc = "Enable expose the precise backlog stats.\n"
-                + " Set false to use published counter and consumed counter to calculate,\n"
-                + " this would be more efficient but may be inaccurate. Default is false."
-    )
-    private boolean exposePreciseBacklogInPrometheus = false;
-
-    @FieldContext(
-        category = CATEGORY_METRICS,
-        doc = "Time in milliseconds that metrics endpoint would time out. Default is 30s.\n"
-                + " Increase it if there are a lot of topics to expose topic-level metrics.\n"
-                + " Set it to 0 to disable timeout."
-    )
-    private long metricsServletTimeoutMs = 30000;
-
-    @FieldContext(
-            category = CATEGORY_METRICS,
-            doc = "Enable expose the backlog size for each subscription when generating stats.\n"
-                    + " Locking is used for fetching the status so default to false."
-    )
-    private boolean exposeSubscriptionBacklogSizeInPrometheus = false;
-
-    @FieldContext(
-            category = CATEGORY_METRICS,
-            doc = "Enable splitting topic and partition label in Prometheus.\n"
-                    + " If enabled, a topic name will split into 2 parts, one is topic name without partition index,\n"
-                    + " another one is partition index, e.g. (topic=xxx, partition=0).\n"
-                    + " If the topic is a non-partitioned topic, -1 will be used for the partition index.\n"
-                    + " If disabled, one label to represent the topic and partition, e.g. (topic=xxx-partition-0)\n"
-                    + " Default is false."
-    )
-    private boolean splitTopicAndPartitionLabelInPrometheus = false;
-
-    @FieldContext(
-            dynamic = true,
-            category = CATEGORY_METRICS,
-            doc = "Enable expose the broker bundles metrics."
-    )
-    private boolean exposeBundlesMetricsInPrometheus = false;
-
     /**** --- Functions. --- ****/
     @FieldContext(
         category = CATEGORY_FUNCTIONS,
@@ -2919,28 +2838,6 @@ public class ServiceConfiguration implements PulsarConfiguration {
     )
     private boolean functionsWorkerEnablePackageManagement = false;
 
-    /**** --- Broker Web Stats. --- ****/
-    @FieldContext(
-        category = CATEGORY_METRICS,
-        doc = "If true, export publisher stats when returning topics stats from the admin rest api"
-    )
-    private boolean exposePublisherStats = true;
-    @FieldContext(
-        category = CATEGORY_METRICS,
-        minValue = 1,
-        doc = "Stats update frequency in seconds"
-    )
-    private int statsUpdateFrequencyInSecs = 60;
-    @FieldContext(
-        category = CATEGORY_METRICS,
-        doc = "Stats update initial delay in seconds"
-    )
-    private int statsUpdateInitialDelayInSecs = 60;
-    @FieldContext(
-        category = CATEGORY_METRICS,
-        doc = "If true, aggregate publisher stats of PartitionedTopicStats by producerName"
-    )
-    private boolean aggregatePublisherStatsByProducerName = false;
 
     /**** --- Ledger Offloading. --- ****/
     /****
@@ -3401,20 +3298,6 @@ public class ServiceConfiguration implements PulsarConfiguration {
             // Note: chroot is not settable by using 'zookeeperServers' config.
             return "metadata-store:" + getMetadataStoreUrl() + BookKeeperConstants.DEFAULT_ZK_LEDGERS_ROOT_PATH;
         }
-    }
-
-    public Object getProperty(String key) {
-        return properties.get(key);
-    }
-
-    @Override
-    public Properties getProperties() {
-        return properties;
-    }
-
-    @Override
-    public void setProperties(Properties properties) {
-        this.properties = properties;
     }
 
     public boolean isDefaultTopicTypePartitioned() {
